+++ src/services/MailService.ts
@@ -2,10 +2,11 @@ import { ConfigData } from '@/configuration';
 import logger from '@/helper/logger';
 import { PlatformMessage } from '@/interfaces/PlatformMessage';
 import QueueManager, { JobData } from './QueueManager';
-import { Job, JobId } from 'bull';
+import { Job, JobId, Queue } from 'bull';
 import Mail from '@/interfaces/Mail';
-import { MailTransport } from './MailTransport';
+import { MailTransport, MailError } from './MailTransport';
 import { SentMessageInfo } from 'nodemailer';
+import FailedJobModel from '@/models/failedJobs';
 
 const SERVICE_TYPE = 'mail';
 
@@ -83,7 +84,9 @@ export default class MailService {
 		logger.debug(
 			`[queue] ${job.queue.name} Processing job id: ${job.id}, messageId: ${messageId}, receiver: ${receiver}`
 		);
-		return this.getTransport(SERVICE_TYPE, platformId).deliver(message as Mail);
+		return this.getTransport(SERVICE_TYPE, platformId)
+			.deliver(message as Mail)
+			.catch(this.errorHandler(job));
 	}
 
 	// TODO refactor to a TransportManager?
@@ -94,5 +97,63 @@ export default class MailService {
 		const transport = transports[randPos];
 
 		return transport;
+	}
+
+	private errorHandler(job: Job<JobData>) {
+		// eslint-disable-next-line @typescript-eslint/no-explicit-any
+		return (error: MailError) => {
+			// TODO refactor error handling for better unit testing
+			// ionos email provider is used for live systems
+			// https://www.ionos.de/hilfe/e-mail/postmaster/smtp-fehlermeldungen-der-11-ionos-mailserver/
+
+			// TODO: make backup jobs available through a route
+
+			// remove jobs with invalid DNS
+			if (error.responseCode >= 550) {
+				this.backupFailedJob(job, error);
+			} else if (
+				error.responseCode === 421 &&
+				error.message.includes('421 Rate limit reached. Please try again later')
+			) {
+				// const pauseDelay = 3 * 60 * 1000;
+				const pauseDelay = 2000;
+				// TODO: eskalation. send email to admin + send to sentry
+				logger.error(
+					`[Critical Error] Rate limit reached, pausing queue ${job.queue.name} for ${pauseDelay} ms`,
+					error
+				);
+				this.pauseQueue(job.queue, pauseDelay);
+			} else if (error.responseCode === 421 && error.message.includes('421 Reject due to policy violations')) {
+				logger.error(
+					`[Critical Error] E-mail account on queue ${job.queue.name} is blocked, please contact Ionos`,
+					error
+				);
+			}
+			throw error;
+		};
+	}
+
+	private async backupFailedJob(job: Job<JobData>, error: Error): Promise<void> {
+		try {
+			const doc = await FailedJobModel.create({
+				receiver: job.data.receiver,
+				jobId: job.id,
+				data: job.data,
+				error: error,
+			});
+			logger.debug('Failed job saved', { id: doc.id, receiver: job.data.receiver });
+		} catch (error) {
+			logger.error('Could not save failed job', job);
+		}
+	}
+
+	private pauseQueue(queue: Queue, delay: number) {
+		logger.debug(`Pausing queue ${queue.name}`);
+		queue.pause().then(() => {
+			setTimeout(() => {
+				logger.debug(`Resuming queue ${queue.name}`);
+				queue.resume();
+			}, delay);
+		});
 	}
 }