+++ src/services/MailTransport.ts
@@ -1,6 +1,5 @@
 import { ConfigData } from '@/configuration';
 import nodemailer, { SentMessageInfo } from 'nodemailer';
-import util from 'util';
 import Mail, { Attachment } from '@/interfaces/Mail';
 import logger from '@/helper/logger';
 
@@ -12,10 +11,17 @@ export interface MessageTransport<T> {
 }
 
 export type MessageTransportStatus = {
-	lastSuccessAt: Date | undefined;
-	lastErrorAt: Date | undefined;
-	lastError: Error | undefined;
-	unavailableSince: Date | undefined;
+	lastSuccessAt: Date | null;
+	lastErrorAt: Date | null;
+	lastError: Error | null;
+	unavailableSince: Date | null;
+};
+
+export type MailError = {
+	name: string;
+	message: string;
+	stack?: string;
+	responseCode: number;
 };
 
 export class MailTransport implements MessageTransport<Mail> {
@@ -23,10 +29,10 @@ export class MailTransport implements MessageTransport<Mail> {
 	private msgDefaults: ConfigData;
 
 	_status: MessageTransportStatus = {
-		lastSuccessAt: undefined,
-		lastErrorAt: undefined,
-		lastError: undefined,
-		unavailableSince: undefined,
+		lastSuccessAt: null,
+		lastErrorAt: null,
+		lastError: null,
+		unavailableSince: null,
 	};
 
 	/**
@@ -79,11 +85,9 @@ export class MailTransport implements MessageTransport<Mail> {
 			const sentInfo = await this.transporter.sendMail(message);
 
 			this._status.lastSuccessAt = new Date();
-			this._status.unavailableSince = undefined;
+			this._status.unavailableSince = null;
 
-			logger.debug(
-				`[transport] Message delivered on ${this.serviceType}/${this.platformId}: ${util.inspect(sentInfo)}`
-			);
+			logger.debug(`[transport] Message delivered on ${this.serviceType}/${this.platformId}`, sentInfo);
 			return sentInfo;
 		} catch (error) {
 			this._status.lastErrorAt = new Date();