+++ src/services/QueueManager.ts
@@ -1,8 +1,9 @@
 import { ConfigData } from '@/configuration';
-import Bull, { Queue, ProcessPromiseFunction, JobOptions, JobCounts } from 'bull';
+import Bull, { Queue, ProcessPromiseFunction, JobOptions, JobCounts, JobId, Job } from 'bull';
 import logger from '@/helper/logger';
 import { PlatformMessage } from '@/interfaces/PlatformMessage';
 import { RedisOptions } from 'ioredis';
+import { ValidationError } from '@/errors';
 
 export type JobData = {
 	serviceType: string;
@@ -13,7 +14,11 @@ export type JobData = {
 };
 
 export default class QueueManager {
-	private queues: Queue[] = [];
+	private _queues: Queue[] = [];
+
+	get queues(): Queue[] {
+		return this._queues;
+	}
 
 	/**
 	 *
@@ -24,7 +29,7 @@ export default class QueueManager {
 	createQueue(serviceType: string, platformId: string, config: ConfigData): Queue {
 		const name = this.getQueueName(serviceType, platformId);
 
-		logger.debug(`[queue] initialize service queue ${name} with config:\n`, config);
+		logger.debug(`[queue] Initializing service queue ${name}`);
 
 		const queue = new Bull<JobData>(name, {
 			prefix: config.defaults.prefix,
@@ -35,20 +40,11 @@ export default class QueueManager {
 			},
 		});
 
-		// TODO Do we need a ready event? In Bull this is only provided via Queue.isReady()
+		// resume on restart
+		queue.resume();
 
-		queue.on('error', (err) => {
-			logger.error(`[queue] ${queue.name} failed with error ${err.message}`);
-		});
-
-		queue.on('failed', (job, err) => {
-			logger.error(`[queue] ${queue.name} Job ${job.id} failed with error ${err.message}`);
-			// TODO log 'retrying'
-		});
-
-		queue.on('stalled', (job) => {
-			logger.warn(`[queue] ${queue.name}: Job ${job.id} stalled and will be reprocessed`);
-		});
+		this.attachQueueEventHandlers(queue);
+		this.attachRedisEventHandlers(queue);
 
 		// TODO check if there's already a queue with the same name
 		this.queues.push(queue);
@@ -65,6 +61,8 @@ export default class QueueManager {
 	async startWorker(serviceType: string, platformId: string, callback: ProcessPromiseFunction<JobData>): Promise<void> {
 		const queue = this.findQueue(serviceType, platformId);
 
+		// this is important to ensure during startup that we are initially connected to redis
+		// otherwise the reconnect doesn't work
 		logger.info(`[queue] Checking if ${queue.name} is ready`);
 		await queue.isReady();
 		logger.info(`[queue] Ok ${queue.name} is ready`);
@@ -77,10 +75,13 @@ export default class QueueManager {
 	 *
 	 * @param data
 	 */
-	async addJob(data: JobData): Promise<void> {
+	async addJob(data: JobData): Promise<JobId> {
 		const queue = this.findQueue(data.serviceType, data.platformId);
-		await queue.add(data);
-		logger.debug(`[queue] ${queue.name} job added:`, { data });
+		const job = await queue.add(data);
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
+		const { message, ...logData } = data;
+		logger.debug(`[queue] ${queue.name} Job id=${job.id} added:`, logData);
+		return job.id;
 	}
 
 	/**
@@ -115,8 +116,10 @@ export default class QueueManager {
 	private findQueue(serviceType: string, platformId: string): Queue {
 		const queue = this.queues.find((q) => q.name === this.getQueueName(serviceType, platformId));
 		if (!queue) {
-			// TODO throw a more specific error that can be evaluated to a status code (422)
-			throw new Error(`Could not find queue with platformId='${platformId}' and serviceType='${serviceType}'`);
+			throw new ValidationError(
+				`Could not find queue with platformId='${platformId}' and serviceType='${serviceType}'`,
+				[]
+			);
 		}
 		return queue;
 	}
@@ -127,18 +130,15 @@ export default class QueueManager {
 
 	private getRedisOptions(config: ConfigData): RedisOptions {
 		const options = config.defaults.redis;
-
-		// https://www.npmjs.com/package/ioredis#auto-reconnect
+		// https://github.com/luin/ioredis#auto-reconnect
+		// never retry failed requests
+		// NOTE: This is per request!
+		options.maxRetriesPerRequest = 0;
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
 		options.retryStrategy = (times: number) => {
-			// one hour every 10 sec retry
-			if (times >= config.defaults.redisRetryAttempts) {
-				logger.error('[critical] Unable to connect to the Redis server - Notification Service is going to exit!');
-				process.exit(1);
-			}
-			logger.error('Unable to connect to the Redis server ..retry', { times });
-			return 10000;
+			return 5000; // reconnect after 5 seconds
 		};
-		logger.debug('redis config: ', options);
+		logger.debug('[queue] Redis config: ', options);
 		return options;
 	}
 
@@ -153,5 +153,81 @@ export default class QueueManager {
 			removeOnComplete: config.defaults.removeOnSuccess,
 			removeOnFail: config.defaults.removeOnFailure,
 		};
+	}
+
+	private attachQueueEventHandlers(queue: Queue) {
+		// https://github.com/OptimalBits/bull/blob/develop/REFERENCE.md#events
+		queue.on('error', (error) => {
+			logger.error(`[queue] ${queue.name} ${error}`);
+		});
+		queue.on('waiting', (jobId) => {
+			// A Job is waiting to be processed as soon as a worker is idling.
+			logger.debug(`[queue] ${queue.name} Job id=${jobId} waiting`);
+		});
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
+		queue.on('active', (job, jobPromise) => {
+			// A job has started. You can use `jobPromise.cancel()`` to abort it.
+			logger.debug(`[queue] ${queue.name} Job id=${job.id} active`);
+		});
+		queue.on('stalled', (job) => {
+			// A job has been marked as stalled. This is useful for debugging job
+			// workers that crash or pause the event loop.
+			logger.error(`[queue] ${queue.name} Job id=${job.id} stalled`);
+		});
+		queue.on('progress', (job, progress) => {
+			// A job's progress was updated!
+			logger.debug(`[queue] ${queue.name} Job id=${job.id} progress`, progress);
+		});
+		queue.on('completed', (job, result) => {
+			// A job successfully completed with a `result`.
+			logger.debug(`[queue] ${queue.name} Job id=${job.id} completed`, result);
+		});
+		queue.on('failed', (job, error) => {
+			// A job failed with reason `error`!
+			logger.error(`[queue] ${queue.name} Job id=${job.id} failed`, error);
+		});
+		queue.on('paused', () => {
+			// The queue has been paused.
+			logger.debug(`[queue] ${queue.name} paused`);
+		});
+		queue.on('resumed', (job: Job<JobData>) => {
+			// The queue has been resumed.
+			logger.debug(`[queue] ${queue.name} ${job ? 'Job id=' + job.id + ' ' : ''}resumed`);
+		});
+		queue.on('cleaned', (jobs, type) => {
+			// Old jobs have been cleaned from the queue. `jobs` is an  of cleaned
+			// jobs, and `type` is the type of jobs cleaned.
+			logger.debug(`[queue] ${queue.name} Jobs ids=${jobs.map((j) => j.id)} cleaned`, type);
+		});
+		queue.on('drained', () => {
+			// Emitted every time the queue has processed all the waiting jobs (even if there can be some delayed jobs not yet processed)
+			logger.debug(`[queue] ${queue.name} drained`);
+		});
+		queue.on('removed', (job) => {
+			// A job successfully removed.
+			logger.info(`[queue] ${queue.name} Job id=${job.id} removed`);
+		});
+	}
+
+	private attachRedisEventHandlers(queue: Queue) {
+		// https://github.com/luin/ioredis#connection-events
+		queue.client.on('connect', (...args) => {
+			logger.debug(`[redis] ${queue.name} connect`, args);
+		});
+		queue.client.on('ready', (...args) => {
+			logger.debug(`[redis] ${queue.name} ready`, args);
+		});
+		queue.client.on('error', (...args) => {
+			logger.debug(`[redis] ${queue.name} error`, args);
+		});
+		queue.client.on('close', (...args) => {
+			logger.debug(`[redis] ${queue.name} close`, args);
+		});
+		queue.client.on('reconnecting', (...args) => {
+			logger.debug(`[redis] ${queue.name} reconnecting`, args);
+		});
+		queue.client.on('end', (...args) => {
+			logger.debug(`[redis] ${queue.name} end`, args);
+		});
 	}
 }