+++ CHANGELOG.md
@@ -11,12 +11,19 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
+- Added E-mail error handling
+- Added status routes for queues and transports
 - OPS-1297 - Added Changelog github action
 
 ### Fixed
 
 - ARC-138 fix changelog action
 - added missing package lock file
+- fixed debug start scripts
+
+### Changed
+
+- SC-6832, SC-7568 - New implementation of mail service
 
 # Release 1.2.2